move tests out of editor module
